move test file to proper folder and namespace
