--- src/Tests/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="2.6.3" targetFramework="net451" />
-  <package id="Rx-Core" version="2.2.2" targetFramework="net451" />
-  <package id="Rx-Interfaces" version="2.2.2" targetFramework="net451" />
-  <package id="Rx-Linq" version="2.2.2" targetFramework="net451" />
-  <package id="Rx-Main" version="2.2.2" targetFramework="net451" />
-  <package id="Rx-PlatformServices" version="2.2.2" targetFramework="net451" />
-</packages>
